+++ modules/core/builtin/cpp/Gateway.cpp
@@ -32,6 +32,7 @@
 #include "prefdirBuiltin.hpp"
 #include "GetPreferencesPath.hpp"
 #include "maxNumCompThreadsBuiltin.hpp"
+#include "existBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -53,6 +54,7 @@ static const nlsGateway gateway[] =
     { "nargout", Nelson::CoreGateway::nargoutBuiltin, 1, 1 },
     { "prefdir", Nelson::CoreGateway::prefdirBuiltin, 1, 0 },
     { "maxNumCompThreads", Nelson::CoreGateway::maxNumCompThreadsBuiltin, 1, -1},
+	{ "exist", Nelson::CoreGateway::existBuiltin, 1, 2 },
 };
 //=============================================================================
 NLSGATEWAYNAME()