+++ CHANGELOG.md
@@ -1,9 +1,17 @@
+
 Features:
 ---------
+
+* add 'exist' function to check existence of variable, builtin, function, file or directory.
 
 * add 'f2c' function to convert fortran code to C from Nelson
 
 * add Fortran 2 C converter library (based on libf2c forked)
+
+Compilation:
+---------
+
+  SLICOT used for control functions.
 
 
 ## 0.1.9 alpha (2017-09-02)